[Vis Tools] Remove /tools/visualization (#6170)
This PR removes /tools/visualization, leaving only the redirect to the
new standardized /tools/map, /tools/scatter, /tools/timeline experience.

## Description

This PR cleans up the legacy `/tools/visualization` route and its
associated UI components. It converts the page into a lightweight
client-side redirector that sends users to the standardized
`/tools/map`, `/tools/scatter`, or `/tools/timeline` pages based on the
URL hash parameters.

This is the first part of the cleanup, focusing **only** on the
tools/visualization path. Updates to the individual tool pages and their
tests will be handled in separate atomic PRs.


## Changes
### 🔧 Server-side Routing & Templates
- **`server/routes/tools/html.py`**
- Simplified the `/tools/visualization` route handler to render the
template without server-side data fetching.
- **`server/templates/tools/visualization.html`**
  - Stripped unused styles and Maps API integrations.
### ⚛️ Client-side App & Redirection
- **`static/js/apps/visualization/main.ts`**
- Updated to perform immediate client-side redirection based on the URL
hash.
- **[DELETE] `static/js/apps/visualization/app.tsx`** and associated UI
components (except for redirect utilities).
- **[DELETE] `static/css/tools/visualization.scss`** and removed its
entry from `webpack.config.js`.
### 🛠️ Shared Utilities
- **`static/js/utils/app/visualization_utils.tsx`**
- Decoupled from the deleted `app_context.ts` by defining
`ContextStatVar` and `DisplayOptions` interfaces locally. This unblocks
chart tiles that still share these utilities without pulling in old
visualization app context.


## Testing Strategy
### Manual Verification
1. Navigate to `/tools/visualization#visType=timeline` (or map/scatter)
and verify that it immediately redirects to the new `/tools/timeline`
page (or map/scatter) with the hash parameters intact.
2. Verify that the build completes successfully without `app_context`
missing errors.
3. Verify that even with standardizd_vis_tool flag set to False,
redirect still happens as intended
4. Verify that webdriver, python, and client tests pass locally.
