+++ server/templates/tools/visualization.html
@@ -16,24 +16,15 @@
 
 {%- extends BASE_HTML -%}
 
-{% set is_hide_full_footer = true %}
 {% set title = "Tools" %}
 
 {% block head %}
-  <link rel="stylesheet" href={{url_for('static', filename='css/visualization.min.css')}} >
-  <link rel="stylesheet" href="https://fonts.googleapis.com/icon?family=Material+Icons">
 {% endblock %}
 
 {% block content %}
-  <div id="metadata" class="d-none" data-map-footer="{{ config['MAP_TOOL_FOOTER'] }}"></div>
   <div id="main-pane"></div>
-  <script>
-    infoConfig = {{info_json|tojson|safe}};
-    globalThis.minStatVarGeoCoverage = {{ config['MIN_STAT_VAR_GEO_COVERAGE'] | int(1) }};
-  </script>
 {% endblock %}
 
 {% block footer %}
   <script src={{url_for('static', filename='visualization.js', t=config['VERSION'])}}></script>
-  <script src="https://maps.googleapis.com/maps/api/js?key={{maps_api_key}}&libraries=places&loading=async" async defer></script>
 {% endblock %}
