--- server/webdriver/tests/vis_scatter_test.py
@@ -1,20 +1,0 @@
-# Copyright 2023 Google LLC
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#      http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-from server.webdriver.base_dc_webdriver import BaseDcWebdriverTest
-from server.webdriver.shared_tests.vis_scatter_test import VisScatterTestMixin
-
-
-class TestVisScatter(VisScatterTestMixin, BaseDcWebdriverTest):
-  """Class to test scatter visualization page. Tests come from VisScatterTestMixin."""
