--- static/js/apps/visualization/app.tsx
@@ -1,97 +1,0 @@
-/**
- * Copyright 2023 Google LLC
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-/**
- * Main component for Visualization Tool.
- */
-
-import { ThemeProvider } from "@emotion/react";
-import _ from "lodash";
-import React, { useContext, useEffect, useState } from "react";
-import { RawIntlProvider } from "react-intl";
-
-import { Spinner } from "../../components/spinner";
-import { intl } from "../../i18n/i18n";
-import { RankingUnitUrlFuncContext } from "../../shared/context";
-import theme from "../../theme/theme";
-import { isSelectionComplete } from "../../utils/app/visualization_utils";
-import { AppContext, AppContextProvider } from "./app_context";
-import { Chart } from "./chart";
-import { Info } from "./info";
-import { SelectedOptions } from "./selected_options";
-import { SelectorPane } from "./selector_pane";
-import { VisTypeSelector } from "./vis_type_selector";
-
-export function App(): JSX.Element {
-  return (
-    <ThemeProvider theme={theme}>
-      <RawIntlProvider value={intl}>
-        <AppContextProvider>
-          <div className="visualization-app">
-            <VisTypeSelector />
-            <MainPane />
-          </div>
-        </AppContextProvider>
-      </RawIntlProvider>
-    </ThemeProvider>
-  );
-}
-
-function MainPane(): JSX.Element {
-  const { visType, places, statVars, enclosedPlaceType, isContextLoading } =
-    useContext(AppContext);
-  const [showInfo, setShowInfo] = useState(
-    _.isEmpty(places) && _.isEmpty(statVars) && _.isEmpty(enclosedPlaceType)
-  );
-
-  useEffect(() => {
-    // When places, statVars, or enclosedPlaceType changes, check if we should
-    // remove the info page.
-    if (
-      !_.isEmpty(places) ||
-      !_.isEmpty(statVars) ||
-      !_.isEmpty(enclosedPlaceType)
-    ) {
-      setShowInfo(false);
-    }
-  }, [places, statVars, enclosedPlaceType]);
-
-  if (isContextLoading) {
-    return (
-      <div className="context-loading-spinner">
-        <Spinner isOpen={true} />
-      </div>
-    );
-  }
-
-  return (
-    <>
-      {showInfo && <Info onStartClicked={(): void => setShowInfo(false)} />}
-      {!showInfo && (
-        <>
-          <SelectedOptions />
-          {isSelectionComplete(visType, places, enclosedPlaceType, statVars) ? (
-            <RankingUnitUrlFuncContext.Provider value={(): null => null}>
-              <Chart />
-            </RankingUnitUrlFuncContext.Provider>
-          ) : (
-            <SelectorPane />
-          )}
-        </>
-      )}
-    </>
-  );
-}
