+++ static/js/apps/visualization/main.ts
@@ -18,29 +18,8 @@
  * Entrypoint for Visualization Tool.
  */
 
-import React from "react";
-import ReactDOM from "react-dom";
-
-import { loadLocaleData } from "../../i18n/i18n";
-import {
-  isFeatureEnabled,
-  STANDARDIZED_VIS_TOOL_FEATURE_FLAG,
-} from "../../shared/feature_flags/util";
-import { App } from "./app";
 import { getStandardizedToolUrl } from "./redirect_utils";
 
 window.addEventListener("load", (): void => {
-  loadLocaleData("en", [import("../../i18n/compiled-lang/en/units.json")]).then(
-    () => {
-      if (isFeatureEnabled(STANDARDIZED_VIS_TOOL_FEATURE_FLAG)) {
-        // If standardized vis tool flag is on, redirect to the old tools
-        window.location.href = getStandardizedToolUrl();
-      } else {
-        ReactDOM.render(
-          React.createElement(App),
-          document.getElementById("main-pane")
-        );
-      }
-    }
-  );
+  window.location.href = getStandardizedToolUrl();
 });
