+++ static/webpack.config.js
@@ -157,10 +157,7 @@ const config = {
       __dirname + "/js/apps/homepage/main_custom_dc.ts",
       __dirname + "/css/homepage.scss",
     ],
-    visualization: [
-      __dirname + "/js/apps/visualization/main.ts",
-      __dirname + "/css/tools/visualization.scss",
-    ],
+    visualization: [__dirname + "/js/apps/visualization/main.ts"],
   },
   output: {
     path: path.resolve(__dirname, "../") + "/server/dist",
